Extract nested-list schema queries to SchemaAnchorInspector
findAnchorColumn, hasAnchor, hasNestedList, allArrayScopes (instance
and static), and immediateScope moved from SpreadsheetSchema to a new
schema.internal.SchemaAnchorInspector helper. Mirrors the existing
BackWriteProjection pattern — nested-list-specific lookups live next to
NestedAnchorValidator under schema.internal, with the same "Not part of
the public API" javadoc disclaimer.

SpreadsheetSchema retains only its core column data and the hot-path
resolve / resolveArray cache. Callers (SheetParser, RecordTreeBuffer,
NestedAnchorValidator) updated to invoke SchemaAnchorInspector. Six
methods drop off the public API surface; behaviour unchanged.
